doing schema work
the work on the database might be postponded temporarily as the database which I had done my research on RealmDB is deprecated. A new alternative that supports the level of encryption we require will be researched. and found will be leaving this branch and working on Authentication. The scope of authentication will be to get our key generation and phrase encryption working.